LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/orthogonal-wm-container-query-expected.txt

XX

PASS Initial non-orthogonal width
FAIL Orthogonal width assert_equals: expected 50 but got 100
